adding archived and program messaging in course header
