+++ src/components/course/CourseAssociatedPrograms.jsx
@@ -2,38 +2,7 @@ import React, { useContext } from 'react';
 
 import { CourseContext } from './CourseContextProvider';
 
-import MicroMastersSvgIcon from '../../assets/icons/micromasters.svg';
-import ProfessionalSvgIcon from '../../assets/icons/professional.svg';
-import VerifiedSvgIcon from '../../assets/icons/verified.svg';
-import XSeriesSvgIcon from '../../assets/icons/xseries.svg';
-import CreditSvgIcon from '../../assets/icons/credit.svg';
-
-function formatProgramType(programType) {
-  switch (programType.toLowerCase()) {
-    case 'micromasters':
-    case 'microbachelors':
-      return <>{programType}<sup>&reg;</sup> Program</>;
-    case 'masters':
-      return 'Master\'s';
-    default:
-      return programType;
-  }
-}
-
-function getProgramIcon(type) {
-  switch (type) {
-    case 'XSeries':
-      return XSeriesSvgIcon;
-    case 'Professional Certificate':
-      return ProfessionalSvgIcon;
-    case 'MicroMasters':
-      return MicroMastersSvgIcon;
-    case 'Credit':
-      return CreditSvgIcon;
-    default:
-      return VerifiedSvgIcon;
-  }
-}
+import { getProgramIcon, formatProgramType } from './data/utils';
 
 export default function CourseAssociatedPrograms() {
   const { state } = useContext(CourseContext);
