--- src/components/course/data/utils.js
@@ -1,36 +1,0 @@
-export function hasCourseStarted(start) {
-  const today = new Date();
-  const startDate = new Date(start);
-  return startDate && today >= startDate;
-}
-
-export function isUserEnrolledInCourse({ userEnrollments, key }) {
-  return userEnrollments.some(({ courseDetails: { courseId } }) => courseId === key);
-}
-
-export function isUserEntitledForCourse({ userEntitlements, courseUuid }) {
-  return userEntitlements.some(({ courseUuid: uuid }) => uuid === courseUuid);
-}
-
-export function weeksRemainingUntilEnd(courseRun) {
-  const today = new Date();
-  const end = new Date(courseRun.end);
-  const secondsDifference = Math.abs(end - today) / 1000;
-  const days = Math.floor(secondsDifference / 86400);
-  return Math.floor(days / 7);
-}
-
-export function hasTimeToComplete(courseRun) {
-  return courseRun.weeksToComplete <= weeksRemainingUntilEnd(courseRun);
-}
-
-export function isArchived(courseRun) {
-  if (courseRun.availability) {
-    return courseRun.availability.toLowerCase() === 'archived';
-  }
-  return false;
-}
-
-export function isCourseSelfPaced(pacingType) {
-  return pacingType === 'self_paced';
-}
