+++ src/components/course/styles/CourseHeader.scss
@@ -8,7 +8,7 @@
       &:focus:before {
         @extend .rounded-0;
       }
-      .enroll-label {
+      .enroll-btn-label {
         font-size: 1.125rem;
       }
     }
